Create files in app's "files" directory (Issue #30)
When relative file paths are specified for the FileAppender, the files are
created relative to the root directory, which is typically non-writable in
Android, leading to a FileNotFoundException when the appender tries to open the
file for writing. The workaround is to use absolute paths, pointing to
somewhere writable for the app.

The default destination for writing out files in Android (via openFileOutput())
is the app's "files" directory (getFilesDir()), where the app always has
write-permission -- all things being equal.

This changes the FileAppender so that its relative path (if any) is prefixed
with the app's "files" directory path (converting it to an absolute path to a
writable location).