Fix possible race condition in SMTPAppender_GreenTest
testCustomEvaluator() was failing as it was seeing the expected
result from the previous test. This failure was only noticed on
Jenkins but not on my development machine.

This change attempts to wait for an email to be received before
trying to read the queue, which could contain messages from
previous tests.