Fix error when child config has enclosing tag (#66)
A configuration error occurred when an included configuration was
enclosed in a <configuration> or <included> tag. The closing tag
of that enclosure was not detected and thus not removed from the
XML doc, which caused it to propagate to the parent config, leading
to an XML parsing error. This was seen as a confusing error to the
user:

   Failed to instantiate [ch.qos.logback.classic.LoggerContext]

The problem was that the SaxEvent's qName was not populated for
the closing tag of the included config. However, the localName
was properly set. The IncludeAction now checks the qName and falls
back to localName if needed.

Fixes Issue #66